+++ core-ng/src/main/java/core/framework/internal/web/request/RequestParser.java
@@ -70,6 +70,7 @@ public final class RequestParser {
             String contentType = headers.getFirst(Headers.CONTENT_TYPE);
             request.contentType = contentType == null ? null : ContentType.parse(contentType);
             parseBody(request, exchange);
+            actionLog.stats.put("http_content_length", (double) exchange.getRequestContentLength());
         }
     }