+++ core-ng/src/main/java/core/framework/log/message/ActionLogMessage.java
@@ -34,8 +34,6 @@ public class ActionLogMessage {
     public String errorMessage;
     @Property(name = "elapsed")
     public Long elapsed;
-    @Property(name = "cpu_time")
-    public Long cpuTime;
     @Property(name = "context")
     public Map<String, List<String>> context;
     @Property(name = "stats")